Test Page & Calendar
Edited test page to be able to list the events on the secondary calendar and add a new event. Server stuff is mainly for keeping who's signed in and getting certain info to interact with the api. .env's should also have a session secret that whoever is using it can generate themselves which is reflected in the sample. I don't wholly understand it though